requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
telethon==1.38.1
python-dotenv==1.0.1
aiosqlite==0.20.0
openpyxl==3.1.5
